small code refactor
