+++ bin/fetch-currently-reading.js
@@ -1,34 +1,10 @@
 const fs = require("fs");
-const { request } = require("undici");
-const { XMLParser } = require("fast-xml-parser");
-const parser = new XMLParser();
-
-const READING_FILE_PATH = "./_data/reading.json";
-const READING_RSS = "https://oku.club/rss/collection/SmX9F";
-
-async function fetchCurrentlyReading() {
-  const { statusCode, body } = await request(READING_RSS);
-  if (statusCode >= 400) {
-    throw new Error(`[shelflife] failed due to status: ${statusCode}`);
-  }
-
-  const rawXML = await body.text();
-  const books = parser.parse(rawXML).rss.channel.item;
-  const cleanBooks = books.map((bk) => {
-    return {
-      title: bk.title,
-      author: bk["dc:creator"],
-      link: bk.link,
-      started: bk.pubDate,
-    };
-  });
-  return cleanBooks;
-}
+const { constants, fetchBooksFromRSS } = require("./shelf.js");
 
 async function possiblyUpdateReadingFile(books) {
-  if (!fs.existsSync(READING_FILE_PATH)) {
+  if (!fs.existsSync(constants.READING_FILE_PATH)) {
     await fs.promises.writeFile(
-      READING_FILE_PATH,
+      constants.READING_FILE_PATH,
       JSON.stringify(books, null, 4),
       {
         encoding: "utf8",
@@ -37,14 +13,16 @@ async function possiblyUpdateReadingFile(books) {
     return;
   }
 
-  const existingReading = await fs.promises.readFile(READING_FILE_PATH);
+  const existingReading = await fs.promises.readFile(
+    constants.READING_FILE_PATH,
+  );
   if (JSON.stringify(books, null, 4) === existingReading.toString()) {
     console.log(`[shelflife] no change, leaving file as is.`);
     return;
   }
 
   await fs.promises.writeFile(
-    READING_FILE_PATH,
+    constants.READING_FILE_PATH,
     JSON.stringify(books, null, 4),
     { encoding: "utf8" },
   );
@@ -52,7 +30,7 @@ async function possiblyUpdateReadingFile(books) {
 }
 
 async function main() {
-  const reading = await fetchCurrentlyReading();
+  const reading = await fetchBooksFromRSS(constants.READING_RSS);
   await possiblyUpdateReadingFile(reading);
 }
